Update main.cpp
Added the use of pointers so the new balance and savings is the value that is being used when doing desposits and withdrawals.  (previous code would show the new balance/savings with the overdraf fee, however the value was not used with the calaculations)

Known error with the withdrawal of the balance, currently debugging.